refactor(tests): split docker api integration suites
