+++ app/arcbox-docker/tests/api_system.rs
@@ -1,0 +1,73 @@
+//! Docker API integration tests.
+
+#[path = "support/api.rs"]
+mod api_support;
+use api_support::*;
+use http_body_util::BodyExt;
+
+#[tokio::test]
+#[ignore = "requires running daemon with guest dockerd"]
+async fn test_ping() {
+    let (runtime, connector, _tmp) = create_test_runtime().await;
+    let app = create_router(runtime, connector);
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .uri("/_ping")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+}
+
+#[tokio::test]
+#[ignore = "requires running daemon with guest dockerd"]
+async fn test_version() {
+    let (runtime, connector, _tmp) = create_test_runtime().await;
+    let app = create_router(runtime, connector);
+
+    let response = app
+        .oneshot(
+            Request::builder()
+                .uri("/version")
+                .body(Body::empty())
+                .unwrap(),
+        )
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+
+    assert!(json.get("Version").is_some());
+    assert!(json.get("ApiVersion").is_some());
+    assert!(json.get("Os").is_some());
+    assert!(json.get("Arch").is_some());
+}
+
+#[tokio::test]
+#[ignore = "requires running daemon with guest dockerd"]
+async fn test_info() {
+    let (runtime, connector, _tmp) = create_test_runtime().await;
+    let app = create_router(runtime, connector);
+
+    let response = app
+        .oneshot(Request::builder().uri("/info").body(Body::empty()).unwrap())
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), StatusCode::OK);
+
+    let body = response.into_body().collect().await.unwrap().to_bytes();
+    let json: serde_json::Value = serde_json::from_slice(&body).unwrap();
+
+    assert!(json.get("Containers").is_some());
+    assert!(json.get("Images").is_some());
+    assert!(json.get("ServerVersion").is_some());
+}
